Put serializers into separate files (issue #61)
