+++ postpost/api/serializers/__init__.py
@@ -1,0 +1,6 @@
+from api.serializers.attachment import AttachmentSerializer  # noqa: F401
+from api.serializers.platform_settings_field import PlatformSettingsRelatedField  # noqa: F401
+from api.serializers.publication import PublicationSerializer  # noqa: F401
+from api.serializers.telegram_settings import TelegramChannelSettingsSerializer  # noqa: F401,E501
+from api.serializers.user_registration import UserRegistrationSerializer  # noqa: F401
+from api.serializers.vk_settings import VKGroupSettingsSerializer  # noqa: F401
