+++ setup.cfg
@@ -4,6 +4,7 @@ include_trailing_comma = true
 sections = FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 default_section = FIRSTPARTY
 line_length = 99
+not_skip = __init__.py
 
 [flake8]
 
@@ -28,8 +29,6 @@ per-file-ignores =
     # disable __init__ with logic, because import all models or admin interfaces from submodules is very conveniently
     models/__init__.py: Z412,
     admin/__init__.py: Z412,
-    # temporarily ignore too many imports
-    api/serializers.py: Z201,
     ## Global ignores
     # I don't like "One-line docstring should fit on one line with quotes"
     *.py: D200,
